# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IAdminLoginResponse": {
				"properties": {
					"jwt": {
						"type": "string"
					}
				},
				"required": [
					"jwt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAdminConnectionRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IExercisesGroup": {
				"properties": {
					"is_active": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"is_active",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorCode": {
				"enum": [
					404,
					403,
					400,
					410,
					429,
					500
				],
				"type": "number"
			},
			"StructuredErrors": {
				"type": "string",
				"enum": [
					"sql/failed",
					"sql/not-found",
					"validation/failed",
					"auth/missing-email",
					"auth/missing-group-id",
					"auth/unknown-email",
					"auth/missing-magic-link-token",
					"auth/invalid-magic-link-token",
					"auth/missing-header",
					"auth/access-token-expired",
					"auth/invalid-access-token",
					"user/register/user-already-exists",
					"ssh/connexion failed",
					"mysql/connexion failed",
					"internal/unknown"
				]
			},
			"ApiError": {
				"properties": {
					"httpCode": {
						"$ref": "#/components/schemas/ErrorCode"
					},
					"structuredError": {
						"$ref": "#/components/schemas/StructuredErrors"
					},
					"errMessage": {
						"type": "string"
					},
					"errDetails": {}
				},
				"required": [
					"httpCode",
					"structuredError",
					"errMessage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IChangeExerciseGroupStateRequest": {
				"properties": {
					"state": {
						"type": "boolean"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStudentsResults": {
				"properties": {
					"full_name": {
						"type": "string"
					},
					"points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"full_name",
					"points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IExerciseBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"test": {
						"properties": {
							"port": {
								"type": "number",
								"format": "double"
							},
							"password": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"host": {
								"type": "string"
							}
						},
						"required": [
							"port",
							"username",
							"host"
						],
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUserGroupID.Exclude_keyofIUserGroupID.group_id__": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"full_name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"full_name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUserGroupID.group_id_": {
				"$ref": "#/components/schemas/Pick_IUserGroupID.Exclude_keyofIUserGroupID.group_id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUser": {
				"$ref": "#/components/schemas/Omit_IUserGroupID.group_id_"
			},
			"IIndexResponse_IUser_": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IUser"
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"limit",
					"total",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUser.Exclude_keyofIUser.id__": {
				"properties": {
					"email": {
						"type": "string"
					},
					"full_name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"full_name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.id_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserCreate": {
				"$ref": "#/components/schemas/Omit_IUser.id_"
			},
			"IUpdateResponse": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "number",
								"format": "double"
							},
							{
								"type": "string"
							}
						]
					},
					"rows": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IUserCreate_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"full_name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUserUpdate": {
				"$ref": "#/components/schemas/Partial_IUserCreate_"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "jwt",
				"in": "header",
				"authorizationUrl": "http://swagger.io/api/oauth/dialog"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/login": {
			"get": {
				"operationId": "LoginFromMaginLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"redirectTo": {
											"type": "string"
										},
										"renew": {
											"type": "string"
										},
										"access": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"redirectTo",
										"renew",
										"access"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "jwt",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/login": {
			"post": {
				"operationId": "LoginAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAdminLoginResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAdminConnectionRequest"
							}
						}
					}
				}
			}
		},
		"/admin/exercise_list": {
			"get": {
				"operationId": "GetAllExercises",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IExercisesGroup"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"JWTADMIN": []
					}
				],
				"parameters": []
			}
		},
		"/admin/exercise": {
			"post": {
				"operationId": "ChangeExerciseGroupState",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/ApiError"
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"JWTADMIN": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IChangeExerciseGroupStateRequest"
							}
						}
					}
				}
			}
		},
		"/admin/results": {
			"get": {
				"operationId": "GetStudentsResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IStudentsResults"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"JWTADMIN": []
					}
				],
				"parameters": []
			}
		},
		"/production/exercise": {
			"post": {
				"operationId": "RunExercise",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IExerciseBody"
							}
						}
					}
				}
			}
		},
		"/protected/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICreateResponse"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreate"
							}
						}
					}
				}
			}
		},
		"/protected/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IIndexResponse_IUser_"
								}
							}
						}
					}
				},
				"description": "Récupérer une page d'utilisateurs.",
				"security": [],
				"parameters": [
					{
						"description": "La page (zéro-index) à récupérer",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Le nombre d'éléments à récupérer (max 50)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"ok": {
											"type": "boolean"
										}
									},
									"required": [
										"ok"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Créer un nouvel utilisateur",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"get": {
				"operationId": "ReadUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Récupérer une utilisateur avec le ID passé dans le URL",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Mettre à jour un utilisateur avec le ID passé dans le URL",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdate"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUpdateResponse"
								}
							}
						}
					}
				},
				"description": "Supprimer un utilisateur",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}